Remove demo-rollup dep from demo-prover
