+++ examples/demo-stf/Cargo.toml
@@ -22,6 +22,7 @@ tokio = { workspace = true, optional = true }
 hex = { workspace = true }
 tracing = { workspace = true }
 
+sov-stf-runner = { path = "../../full-node/sov-stf-runner" }
 sov-rollup-interface = { path = "../../rollup-interface" }
 sov-cli = { path = "../../module-system/sov-cli", optional = true }
 sov-sequencer-registry = { path = "../../module-system/module-implementations/sov-sequencer-registry" }
@@ -36,7 +37,6 @@ sov-accounts = { path = "../../module-system/module-implementations/sov-accounts
 sov-state = { path = "../../module-system/sov-state" }
 sov-modules-api = { path = "../../module-system/sov-modules-api" }
 sov-sequencer = { path = "../../full-node/sov-sequencer", optional = true }
-sov-stf-runner = { path = "../../full-node/sov-stf-runner", optional = true }
 # Only enable the evm on "experimental" feature
 sov-evm = { path = "../../module-system/module-implementations/sov-evm", optional = true }
 reth-primitives = { workspace = true, optional = true }
@@ -64,7 +64,6 @@ native = [
     "sov-rollup-interface/mocks",
     "sov-modules-stf-template/native",
     "sov-sequencer",
-    "sov-stf-runner/native",
     "clap",
     "serde_json",
     "jsonrpsee",
