+++ examples/demo-stf/src/lib.rs
@@ -10,4 +10,21 @@ pub mod tests;
 #[cfg(feature = "native")]
 pub mod cli;
 
+use runtime::Runtime;
+use sov_modules_api::default_context::ZkDefaultContext;
+use sov_modules_stf_template::AppTemplate;
+use sov_rollup_interface::da::DaVerifier;
 pub use sov_state::ArrayWitness;
+use sov_stf_runner::verifier::StateTransitionVerifier;
+
+/// A verifier for the demo rollup
+pub type AppVerifier<DA, Zk> = StateTransitionVerifier<
+    AppTemplate<
+        ZkDefaultContext,
+        <DA as DaVerifier>::Spec,
+        Zk,
+        Runtime<ZkDefaultContext, <DA as DaVerifier>::Spec>,
+    >,
+    DA,
+    Zk,
+>;
